+++ gradebookng/tool/src/java/org/sakaiproject/gradebookng/tool/panels/SortGradeItemsByGradeItemPanel.java
@@ -49,7 +49,7 @@ public class SortGradeItemsByGradeItemPanel extends BasePanel {
 			@Override
 			protected void populateItem(final ListItem<Assignment> assignmentItem) {
 				final Assignment assignment = assignmentItem.getModelObject();
-				assignmentItem.add(new Label("name", assignment.getName()).setEscapeModelStrings(false));
+				assignmentItem.add(new Label("name", assignment.getName()));
 				assignmentItem.add(new HiddenField<Long>("id",
 						Model.of(assignment.getId())).add(
 								new AttributeModifier("name",
