+++ jobscheduler/scheduler-component-shared/src/java/org/sakaiproject/component/app/scheduler/jobs/samigo/Messages.properties
@@ -1,6 +1,6 @@
 apply=Apply updates
 apply.description=Set to true to update Gradebook item names. The default false only logs a dry-run summary.
 site.id=Site ID
-site.id.description=Optional site/gradebook UID to limit the backfill to one site. Leave blank to scan all sites.
+site.id.description=Optional site/gradebook UID to limit the scan to one site. Recommended for large systems. Leave blank to scan all sites in batches.
 log.changes=Log changed rows
-log.changes.description=Set to true to log each row whose Gradebook title differs from the decoded Samigo title.
+log.changes.description=Set to true to log each Samigo external Gradebook row whose stored title differs from the decoded published assessment title.
